styles and updates
