+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "homepage": "https://keithbillings.com/",
   "dependencies": {
+    "@svgr/webpack": "^7.0.0",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
