+++ src/scss/base/_variables.scss
@@ -1,0 +1,2 @@
+$navbar-height: 40px;
+$mobile-menu-width: 15rem;
