+++ src/scss/components/_companies.scss
@@ -1,6 +1,8 @@
 .companies {
-
-	&__title {}
-	&__container {}
-	&__logo {}
+  &__title {
+  }
+  &__container {
+  }
+  &__logo {
+  }
 }
