+++ src/scss/components/_hero.scss
@@ -1,7 +1,7 @@
 // hero section for home page
 .hero {
   width: 100%;
-  height: calc(100vh - 40px);
+  height: 100vh;
   background: linear-gradient(180deg, $color-primary-dark-0 0%, $color-primary-dark-50 100%);
   background-color: $color-primary;
 
@@ -26,7 +26,7 @@
     background: linear-gradient(180deg, $color-primary-dark-0 0%, $color-primary-dark-50 100%);
   }
 
-  .hero-container {
+  &__container {
     display: grid;
     grid-template-columns: repeat(2, 1fr);
     max-width: 1440px;
@@ -46,82 +46,101 @@
       max-width: 100vw;
       grid-template-columns: repeat(1, 1fr);
     }
+  }
 
-    .hero-text {
-      transition: all 0.3s;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      width: fit-content;
-      padding-bottom: 20%;
-      padding-left: 10%;
-      cursor: default;
+  &__text {
+    transition: all 0.3s;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    width: fit-content;
+    padding-bottom: 20%;
+    padding-left: 10%;
+    cursor: default;
 
-      @include tablet() {
-        position: absolute;
-        top: 15%;
-        left: 5%;
-        padding-left: 0;
-        padding-bottom: 0%;
-      }
-      @include mobile() {
-        position: absolute;
-        top: 5%;
-        left: 5%;
-        padding-left: 0;
-        padding-bottom: 0%;
-      }
+    @include tablet() {
+      position: absolute;
+      top: 15%;
+      left: 5%;
+      padding-left: 0;
+      padding-bottom: 0%;
+    }
+    @include mobile() {
+      position: absolute;
+      top: 5%;
+      left: 5%;
+      padding-left: 0;
+      padding-bottom: 0%;
+    }
+  }
 
-      &__greeting {
-        color: $color-primary-light;
-        font-size: 2.6rem;
-        margin-left: 5px;
-      }
-      &__name {
-        font-size: 6.8rem;
-        font-family: "Cinzel Decorative";
-        text-shadow: 5px 5px 10px $color-warm-black-20;
+  &__greeting {
+    color: $color-primary-light;
+    font-size: 2.6rem;
+    margin-left: 5px;
+  }
 
-        @include mobile() {
-          font-size: 5rem;
-        }
-      }
-      &__title {
-        font-size: 1.8rem;
-        margin-left: 5px;
-        height: 2.7rem;
-      }
+  &__name {
+    font-size: 6.8rem;
+    font-family: "Cinzel Decorative";
+    text-shadow: 5px 5px 10px $color-warm-black-20;
+
+    @include mobile() {
+      font-size: 5rem;
+    }
+  }
+
+  &__title {
+    font-size: 1.8rem;
+    margin-left: 5px;
+    height: 2.7rem;
+  }
+
+  &__image {
+    transition: all 0.3s;
+    height: calc(90vh - 40px);
+    z-index: -1;
+    margin-top: auto;
+    margin-left: auto;
+    position: relative;
+
+    @include slideLeft(-5rem, 2s);
+
+    @include tablet() {
+      height: 100%;
+      max-height: 60vh;
+      position: absolute;
+      bottom: 0;
+      right: 0;
+    }
+    @include mobile() {
+      height: 100%;
+      max-height: 60vh;
+      position: absolute;
+      bottom: 0;
+      right: 0;
     }
 
-    .hero-image {
-      transition: all 0.3s;
-      height: calc(90vh - 40px);
-      z-index: -1;
-      margin-top: auto;
-      margin-left: auto;
-      position: relative;
+    img {
+      height: 100%;
+      filter: opacity(0.95) drop-shadow(5px 5px 10px $color-warm-black-50);
+    }
+  }
 
-      @include slideLeft(-5rem, 2s);
+  &__background-image {
+    position: absolute;
+    z-index: -2;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    filter: grayscale(100%);
 
-      @include tablet() {
-        height: 100%;
-        max-height: 60vh;
-        position: absolute;
-        bottom: 0;
-        right: 0;
-      }
-      @include mobile() {
-        height: 100%;
-        max-height: 60vh;
-        position: absolute;
-        bottom: 0;
-        right: 0;
-      }
-
-      img {
-        height: 100%;
-        filter: opacity(0.95) drop-shadow(5px 5px 10px $color-warm-black-50);
-      }
+    img {
+      width: 100%;
+      height: 100%;
+      opacity: 0.2;
+      object-fit: cover;
     }
   }
 }
