+++ src/scss/main.scss
@@ -1,4 +1,5 @@
 // Base
+@import "./base/variables";
 @import "./colors/colors";
 @import "./typography/typography";
 @import "./base/base";
